normalizing aleth

+++ libweb3jsonrpc/UnixSocketServer.cpp
@@ -54,19 +54,16 @@ fs::path getIpcPathOrDataDir() {
 }
 }  // namespace
 
-UnixDomainSocketServer::UnixDomainSocketServer(string const& _path)
-    : IpcServerBase((fs::path(_path + "/taraxa.ipc"))
+UnixDomainSocketServer::UnixDomainSocketServer(string const& _appId)
+    : IpcServerBase((getIpcPathOrDataDir() / fs::path(_appId + ".ipc"))
                         .string()
                         .substr(0, c_socketPathMaxLength)) {}
 
 UnixDomainSocketServer::~UnixDomainSocketServer() { StopListening(); }
 
 bool UnixDomainSocketServer::StartListening() {
-  printf("%s startListening\n", m_path.c_str());
   if (!m_running) {
-    if (access(m_path.c_str(), F_OK) != -1) {
-      unlink(m_path.c_str());
-    }
+    if (access(m_path.c_str(), F_OK) != -1) unlink(m_path.c_str());
 
     if (access(m_path.c_str(), F_OK) != -1) return false;
